+++ Makefile.dep
@@ -212,6 +212,7 @@ endif
 
 ifneq (,$(filter gnrc_sixlowpan_frag,$(USEMODULE)))
   USEMODULE += gnrc_sixlowpan
+  USEMODULE += gnrc_sixlowpan_frag_rb
   USEMODULE += xtimer
 endif
