+++ sys/net/gnrc/Makefile
@@ -91,6 +91,9 @@ endif
 ifneq (,$(filter gnrc_sixlowpan_frag,$(USEMODULE)))
   DIRS += network_layer/sixlowpan/frag
 endif
+ifneq (,$(filter gnrc_sixlowpan_frag_rb,$(USEMODULE)))
+  DIRS += network_layer/sixlowpan/frag/rb
+endif
 ifneq (,$(filter gnrc_sixlowpan_frag_vrb,$(USEMODULE)))
   DIRS += network_layer/sixlowpan/frag/vrb
 endif
