+++ sys/net/gnrc/network_layer/sixlowpan/frag/rb/Makefile
@@ -1,0 +1,3 @@
+MODULE := gnrc_sixlowpan_frag_rb
+
+include $(RIOTBASE)/Makefile.base
